Perk system
-Perk system. I need to test how the perk works, but I am almost certain, that it's gonna be ok. UI for the PerkSystemMenu is finished, hovers and keyboard control for it are also finished. Everything is fine.
Also I created a special "Blood magic" perk, which gives an access to the blood magic element. It was tested and is working.
-Fixed some bugs with SpellBattleMenu. SpellRange icon wasn't hidden, when the player opened the spell result. Also I accidentally deleted SpellInfoWidget and it caused some issues with UI. Also fixed a crash bug,
when removing a spell element from horizontal box. Also fixed loading of learned spells and the fact, that not unique spells also were written to the LearnedSpellsMenu.