LAST UPDATE
There are a lot, so I am not going to write a detailed review. New items, quests scripting, new items saving system, enemy spawning, a lot of fixes and bug fixes...
This is certainly the last update, I am officially finished with this game. Now it can be finished, has several quests, enemies and a main quest. Amen.